Minor fixes

+++ inc/libasm.h
@@ -7,7 +7,7 @@ typedef struct s_list
   struct s_list *next;
 } t_list;
 
-int ft_strlen(const char* str);
+unsigned long ft_strlen(const char* str);
 char* ft_strcpy(char* dest, const char* src);
 int ft_strcmp (const char* str1, const char* str2);
 int ft_write(int fd, const void *buf, unsigned int count);
